requirements.txt
Pillow
wxPython
hidapi
